+++ src/otp_twilio/models.py
@@ -1,15 +1,10 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-from binascii import unhexlify
 import logging
-import time
 
 from django.core.exceptions import ImproperlyConfigured
 from django.db import models
 from django.utils.encoding import force_text
 
-from django_otp.models import Device
-from django_otp.oath import TOTP
+from django_otp.models import SideChannelDevice, ThrottlingMixin
 from django_otp.util import hex_validator, random_hex
 import requests
 
@@ -19,21 +14,23 @@ from .conf import settings
 logger = logging.getLogger(__name__)
 
 
-def default_key():
+def default_key():  # pragma: no cover
+    """ Obsolete code here for migrations. """
     return force_text(random_hex(20))
 
 
-def key_validator(value):
+def key_validator(value):  # pragma: no cover
+    """ Obsolete code here for migrations. """
     return hex_validator(20)(value)
 
 
-class TwilioSMSDevice(Device):
+class TwilioSMSDevice(ThrottlingMixin, SideChannelDevice):
     """
-    A :class:`~django_otp.models.Device` that delivers codes via the Twilio SMS
-    service. This uses TOTP to generate temporary tokens, which are valid for
-    :setting:`OTP_TWILIO_TOKEN_VALIDITY` seconds. Once a given token has been
-    accepted, it is no longer valid, nor is any other token generated at an
-    earlier time.
+    A :class:`~django_otp.models.SideChannelDevice` that delivers a token via
+    the Twilio SMS service.
+
+    The tokens are valid for :setting:`OTP_TWILIO_TOKEN_VALIDITY` seconds. Once
+    a token has been accepted, it is no longer valid.
 
     .. attribute:: number
 
@@ -44,38 +41,17 @@ class TwilioSMSDevice(Device):
         will try to infer the correct E.164 format if it is not used, but this
         should not be relied upon.
 
-    .. attribute:: key
-
-        *CharField*: The secret key used to generate TOTP tokens.
-
-    .. attribute:: last_t
-
-        *BigIntegerField*: The t value of the latest verified token.
-
     """
     number = models.CharField(
         max_length=30,
         help_text="The mobile number to deliver tokens to (E.164)."
     )
 
-    key = models.CharField(
-        max_length=40,
-        validators=[key_validator],
-        default=default_key,
-        help_text="A random key used to generate tokens (hex-encoded)."
-    )
-
-    last_t = models.BigIntegerField(
-        default=-1,
-        help_text="The t value of the latest verified token. The next token must be at a higher time step."
-    )
-
-    class Meta(Device.Meta):
+    class Meta(SideChannelDevice.Meta):
         verbose_name = "Twilio SMS Device"
 
-    @property
-    def bin_key(self):
-        return unhexlify(self.key.encode())
+    def get_throttle_factor(self):
+        return settings.OTP_TWILIO_THROTTLE_FACTOR
 
     def generate_challenge(self):
         """
@@ -85,16 +61,16 @@ class TwilioSMSDevice(Device):
         :raises: Exception if delivery fails.
 
         """
-        totp = self.totp_obj()
-        token = format(totp.token(), '06d')
-        message = settings.OTP_TWILIO_TOKEN_TEMPLATE.format(token=token)
+        self.generate_token(valid_secs=settings.OTP_TWILIO_TOKEN_VALIDITY)
+
+        message = settings.OTP_TWILIO_TOKEN_TEMPLATE.format(token=self.token)
 
         if settings.OTP_TWILIO_NO_DELIVERY:
             logger.info(message)
         else:
             self._deliver_token(message)
 
-        challenge = settings.OTP_TWILIO_CHALLENGE_MESSAGE.format(token=token)
+        challenge = settings.OTP_TWILIO_CHALLENGE_MESSAGE.format(token=self.token)
 
         return challenge
 
@@ -135,29 +111,15 @@ class TwilioSMSDevice(Device):
             raise ImproperlyConfigured('OTP_TWILIO_FROM must be set to one of your Twilio phone numbers')
 
     def verify_token(self, token):
-        try:
-            token = int(token)
-        except Exception:
-            verified = False
-        else:
-            totp = self.totp_obj()
-            tolerance = settings.OTP_TWILIO_TOKEN_VALIDITY
+        verify_allowed, _ = self.verify_is_allowed()
+        if verify_allowed:
+            verified = super().verify_token(token)
 
-            for offset in range(-tolerance, 1):
-                totp.drift = offset
-                if (totp.t() > self.last_t) and (totp.token() == token):
-                    self.last_t = totp.t()
-                    self.save()
-
-                    verified = True
-                    break
+            if verified:
+                self.throttle_reset()
             else:
-                verified = False
+                self.throttle_increment()
+        else:
+            verified = False
 
         return verified
-
-    def totp_obj(self):
-        totp = TOTP(self.bin_key, step=1)
-        totp.time = time.time()
-
-        return totp