+++ test/test_project/settings.py
@@ -1,7 +1,5 @@
 # django-agent-trust test project
 
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 from os.path import abspath, dirname, join